refactor(orpc): rewrite server as flat module, drop Stripe from auth/context
- Replace factory-based createORPCProcedures with direct module-level
  procedure definitions mirroring apps/web pattern
- Add withBilling middleware for fire-and-forget HyprPay usage ingestion
- Remove Stripe plugin from Better Auth and stripeClient from oRPC context
  to eliminate TS2883 declaration portability errors
- Make core/authentication and core/orpc non-composite (composite: false)
  to avoid circular declaration emit constraints
- Add explicit Procedure<TCurrent> type annotation on exports to fix TS7056
- Require HYPRPAY_API_KEY (non-optional), remove HYPRPAY_BASE_URL
- Delete billable.ts and procedures.ts (superseded by server.ts)
