+++ core/authentication/src/cli-config.ts
@@ -3,22 +3,22 @@ import { createAuth } from "@core/authentication/server";
 import { createDb } from "@core/database/client";
 import { createRedis } from "@core/redis/connection";
 import { createPostHog } from "@core/posthog/server";
-import { createStripeClient } from "@core/stripe";
 import { createResendClient } from "@core/transactional/utils";
+import { createHyprPayClient } from "@montte/hyprpay";
 import { env } from "@core/environment/web";
 
 const db = createDb({ databaseUrl: env.DATABASE_URL });
 const redis = createRedis(env.REDIS_URL);
 const posthog = createPostHog(env.POSTHOG_KEY, env.POSTHOG_HOST);
-const stripeClient = createStripeClient(env.STRIPE_SECRET_KEY);
 const resendClient = createResendClient(env.RESEND_API_KEY);
+const hyprpayClient = createHyprPayClient({ apiKey: env.HYPRPAY_API_KEY });
 
 const auth = createAuth({
    db,
    redis,
    posthog,
-   stripeClient,
    resendClient,
+   hyprpayClient,
    env,
 });
