+++ core/authentication/src/server.ts
@@ -1,5 +1,4 @@
 import { apiKey } from "@better-auth/api-key";
-import { stripe as stripePlugin } from "@better-auth/stripe";
 import { hyprpay } from "@montte/hyprpay/better-auth";
 import { createHyprPayClient, type HyprPayClient } from "@montte/hyprpay";
 import { findMemberByUserId } from "@core/database/repositories/auth-repository";
@@ -21,7 +20,6 @@ import {
    twoFactor,
 } from "better-auth/plugins";
 import { tanstackStartCookies } from "better-auth/tanstack-start";
-import type Stripe from "stripe";
 import { z } from "zod";
 import type { DatabaseInstance } from "@core/database/client";
 import type { PostHog } from "@core/posthog/server";
@@ -65,39 +63,19 @@ export interface CreateAuthDeps {
    db: DatabaseInstance;
    redis: Redis;
    posthog: PostHog;
-   stripeClient: Stripe;
    resendClient: ResendClient;
+   hyprpayClient: HyprPayClient;
    env: {
       BETTER_AUTH_URL?: string;
       BETTER_AUTH_SECRET: string;
       BETTER_AUTH_TRUSTED_ORIGINS: string;
       BETTER_AUTH_GOOGLE_CLIENT_ID: string;
       BETTER_AUTH_GOOGLE_CLIENT_SECRET: string;
-      STRIPE_WEBHOOK_SECRET: string;
-      STRIPE_BOOST_PRICE_ID?: string;
-      STRIPE_SCALE_PRICE_ID?: string;
-      STRIPE_ENTERPRISE_PRICE_ID?: string;
-      HYPRPAY_API_KEY?: string;
-      HYPRPAY_BASE_URL?: string;
    };
 }
 
-export function createHyprPayClientFromEnv(env: {
-   HYPRPAY_API_KEY?: string;
-   HYPRPAY_BASE_URL?: string;
-}): HyprPayClient | null {
-   return env.HYPRPAY_API_KEY
-      ? createHyprPayClient({
-           apiKey: env.HYPRPAY_API_KEY,
-           baseUrl: env.HYPRPAY_BASE_URL,
-        })
-      : null;
-}
-
 export function createAuth(deps: CreateAuthDeps) {
-   const { db, redis, posthog, stripeClient, resendClient, env } = deps;
-
-   const hyprpayClient = createHyprPayClientFromEnv(env);
+   const { db, redis, posthog, resendClient, hyprpayClient, env } = deps;
 
    const auth = betterAuth({
       baseURL: env.BETTER_AUTH_URL,
@@ -375,103 +353,7 @@ export function createAuth(deps: CreateAuthDeps) {
             apiKeyHeaders: ["sdk-api-key", "x-api-key"],
          }),
 
-         stripePlugin({
-            createCustomerOnSignUp: true,
-            stripeClient,
-            stripeWebhookSecret: env.STRIPE_WEBHOOK_SECRET,
-            subscription: {
-               authorizeReference: async ({ user, referenceId }) => {
-                  const membership = await db.query.member.findFirst({
-                     where: (fields, { and, eq }) =>
-                        and(
-                           eq(fields.organizationId, referenceId),
-                           eq(fields.userId, user.id),
-                        ),
-                  });
-                  if (!membership) return false;
-                  return (
-                     membership.role === "owner" || membership.role === "admin"
-                  );
-               },
-               enabled: true,
-               getCheckoutSessionParams: async () => ({
-                  params: {
-                     allow_promotion_codes: true,
-                  },
-               }),
-               plans: [
-                  {
-                     name: "boost",
-                     priceId: env.STRIPE_BOOST_PRICE_ID,
-                  },
-                  {
-                     name: "scale",
-                     priceId: env.STRIPE_SCALE_PRICE_ID,
-                  },
-                  {
-                     name: "enterprise",
-                     priceId: env.STRIPE_ENTERPRISE_PRICE_ID,
-                  },
-               ],
-            },
-            onSubscriptionComplete: async ({
-               subscription,
-               stripeSubscription,
-            }: {
-               subscription: { id: string; plan: string; referenceId: string };
-               stripeSubscription: Stripe.Subscription;
-            }) => {
-               try {
-                  const invoice = await stripeClient.invoices.retrieve(
-                     stripeSubscription.latest_invoice as string,
-                  );
-                  posthog.capture({
-                     distinctId: subscription.referenceId,
-                     event: "subscription_started",
-                     groups: { organization: subscription.referenceId },
-                     properties: {
-                        $currency: invoice.currency.toUpperCase(),
-                        $revenue: invoice.amount_paid / 100,
-                        interval:
-                           stripeSubscription.items.data[0]?.plan.interval,
-                        organization_id: subscription.referenceId,
-                        plan_name: subscription.plan,
-                        subscription_id: subscription.id,
-                     },
-                  });
-               } catch (error) {
-                  logger.error(
-                     { err: error },
-                     "Failed to capture subscription_started event",
-                  );
-               }
-            },
-            onSubscriptionCancel: async ({
-               subscription,
-            }: {
-               subscription: { id: string; plan: string; referenceId: string };
-            }) => {
-               try {
-                  posthog.capture({
-                     distinctId: subscription.referenceId,
-                     event: "subscription_canceled",
-                     groups: { organization: subscription.referenceId },
-                     properties: {
-                        organization_id: subscription.referenceId,
-                        plan_name: subscription.plan,
-                        subscription_id: subscription.id,
-                     },
-                  });
-               } catch (error) {
-                  logger.error(
-                     { err: error },
-                     "Failed to capture subscription_canceled event",
-                  );
-               }
-            },
-         }),
-
-         hyprpay({ client: hyprpayClient ?? undefined }),
+         hyprpay({ client: hyprpayClient }),
          tanstackStartCookies(),
       ],
    });
